Move URL into _types package
