+++ tamr_client/_types/__init__.py
@@ -1,1 +1,2 @@
 from .json import JsonDict
+from .url import URL
