+++ tamr_client/mastering/project.py
@@ -1,8 +1,7 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from tamr_client._types import JsonDict
-from tamr_client.url import URL
+from tamr_client._types import JsonDict, URL
 
 
 @dataclass(frozen=True)
