+++ tamr_client/operation.py
@@ -9,10 +9,9 @@ from typing import Dict, Optional
 import requests
 
 from tamr_client import response
-from tamr_client._types import JsonDict
+from tamr_client._types import JsonDict, URL
 from tamr_client.instance import Instance
 from tamr_client.session import Session
-from tamr_client.url import URL
 
 
 class NotFound(Exception):
